添加抽卡记录设置实现

+++ src/main/java/com/weird/mapper/RollListMapper.java
@@ -13,7 +13,7 @@ public interface RollListMapper {
 
     int insertSelective(RollListModel record);
 
-    RollListModel selectByPrimaryKey(Integer rollId);
+    RollListModel selectByPrimaryKey(long rollId);
 
     int updateByPrimaryKeySelective(RollListModel record);
